+++ CHANGELOG.markdown
@@ -4,3 +4,4 @@
 * Driver for http://lingvo.yandex.ru. Better translations from English to Russian.
 * Language auto detection.
 * DB schema changes: move transcription property to Translation from Word.
+* Version command for inspect information about the program.